flask_cnn/requirements.txt
Flask
joblib
numpy
Pillow
gunicorn
scikit-learn==1.3.0
tensorflow==2.13.0
